Add new api to get order information by id, update logic to get order to prepare order

+++ OrderService/Models/Responses/CreateOrderDetailResponse.cs
@@ -4,5 +4,5 @@ namespace OrderService.Models.Responses;
 
 public class CreateOrderDetailResponse : BaseResponse
 {
-    public int Data { get; set; }
+    public long Data { get; set; }
 }
